third_party/blink/web_tests/fast/dom/shadow/style-sharing-sibling-shadow-expected.txt

CONSOLE WARNING: line 22: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Siblings should only share if their host rules match

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(host).backgroundColor is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
